refactor: my page에서 사용되던 header 컴포넌트를 공통 컴포넌트로 변경
